Fix package directory path in the GAP root
Prior to this fix, the setup seemed to work fine, unless the package
had the file `lib/somename` which coincides with the the name of some
file in the GAP `lib` directory. In this case GAP read `lib/somename`
from the package instead of the GAP library, and did not read the GAP
library file at all, causing various cryptic error messages. This is
exactly what happened in the Example package, because it contains
`lib/files.{gi,gd}` files.

To solve this, we add an extra level to the additional `pkg` directory
setup, so that the extra GAPROOT we point GAP at is guaranteed to only
contain a `pkg` subdirectory (resp. symlink to a directory), but no
other files which could be read accidentally