Don't mention PreloadFile
This field is not used by any package, and its usefulness
seems highly doubtful. I am tempted to say we should remove it
altogether; but at the very least, I suggest that we do not
suggest to package authors to use it, which we'd effectively do
by having it in the example package.