Update and reduce CI.yml, add CI-advanced.yml
Using 32bit GAP is exceptional these days. And for pure GAP packages it is
always never necessary to test on 32bit. So let's not suggest to people to
test it by default!

Developers working on kernel extensions and who also want to support 32bit
can instead look at the "advanced" CI config file CI-advanced.yml (and use
it as their CI.yml).